refactor: 여행 경로 후보 책임 분리
