+++ src/main/java/com/chocobi/leafy/trip/application/TripSegmentService.java
@@ -12,15 +12,10 @@ import com.chocobi.leafy.trip.dto.request.UpdateTripPlaceRequest;
 import com.chocobi.leafy.trip.dto.response.TripPlaceLocationResponse;
 import com.chocobi.leafy.trip.dto.response.TripPlaceResponse;
 import com.chocobi.leafy.trip.dto.TripSegmentDTO;
-import com.chocobi.leafy.trip.infra.TripFindService;
 import com.chocobi.leafy.trip.infra.TripRouteOptionCommandService;
-import com.chocobi.leafy.trip.infra.TripRouteOptionFindService;
 import com.chocobi.leafy.trip.infra.TripSegmentCommandService;
 import com.chocobi.leafy.trip.infra.TripSegmentFindService;
 import com.chocobi.leafy.trip.infra.entity.TripEntity;
-import com.chocobi.leafy.trip.infra.entity.TripPlaceEntity;
-import com.chocobi.leafy.trip.infra.entity.TripRouteOptionEntity;
-import com.chocobi.leafy.trip.infra.entity.TripSegmentEntity;
 import com.chocobi.leafy.trip.vo.TripTransport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,9 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static com.chocobi.leafy.distance.service.DistanceUtils.placeToPoint;
 
@@ -40,114 +33,12 @@ public class TripSegmentService {
     private final TripSegmentFindService tripSegmentFindService;
     private final TripSegmentCommandService tripSegmentCommandService;
     private final TripRouteOptionCommandService tripRouteOptionCommandService;
-    private final TripRouteOptionFindService tripRouteOptionFindService;
-    private final TripFindService tripFindService;
+    private final TripRouteCandidateService tripRouteCandidateService;
     private final CarDistanceService carDistanceService;
     private final TransDistanceService transDistanceService;
     private final TripPlaceService tripPlaceService;
     private final PlaceService placeService;
 
-    @Transactional
-    public void completeTripSegments(Long tripId, String transport) {
-        if (transport == null) throw new IllegalArgumentException("transport가 필요합니다.");
-
-        TripEntity trip = tripFindService.findTrip(tripId);
-        TripRouteOptionEntity selectedRouteOption = tripRouteOptionFindService.findTripRouteOption(tripId, transport);
-        List<TripRouteOptionEntity> routeOptions = tripRouteOptionFindService.findTripRouteOptions(tripId);
-
-        tripRouteOptionCommandService.confirmOnly(selectedRouteOption, routeOptions);
-        trip.clearRouteStale();
-    }
-
-    @Transactional(readOnly = true)
-    public Map<String, Object> getTotalTimeAndCarbon(Long tripId, String transport) {
-        TripRouteOptionEntity routeOption = tripRouteOptionFindService.findTripRouteOption(tripId, transport);
-
-        Map<String, Object> result = new HashMap<>();
-        result.put("totalDuration", routeOption.getTotalDuration());
-        result.put("totalCarbonEmission", routeOption.getTotalCarbonEmission());
-
-        return result;
-    }
-
-    /**
-     * DB에 TripSegment 저장
-     */
-    public void saveTripSegments(List<TripSegmentEntity> tripSegments) {
-        if (tripSegments == null || tripSegments.isEmpty()) return;
-        tripSegmentCommandService.saveAll(tripSegments);
-    }
-
-    private void saveRouteCandidate(Long tripId, List<Section> sections, String transport, List<TripPlaceResponse> tripPlaces) {
-        validateRouteCandidate(sections, tripPlaces);
-
-        TripEntity trip = tripFindService.findTrip(tripId);
-        TripTransport tripTransport = TripTransport.from(transport);
-        deleteRouteCandidate(tripId, tripTransport);
-
-        TripRouteOptionEntity routeOption = tripRouteOptionCommandService.save(createRouteOptionFromSections(
-                trip,
-                tripTransport,
-                sections,
-                false
-        ));
-        saveTripSegments(createTripSegments(routeOption, sections, tripPlaces));
-    }
-
-    private void validateRouteCandidate(List<Section> sections, List<TripPlaceResponse> tripPlaces) {
-        if (tripPlaces == null || tripPlaces.size() < 2) {
-            throw new IllegalArgumentException("여행 장소는 2개 이상 필요합니다.");
-        }
-
-        if (sections == null || sections.isEmpty()) {
-            throw new IllegalArgumentException("저장할 여행 경로 구간이 없습니다.");
-        }
-    }
-
-    private void deleteRouteCandidate(Long tripId, TripTransport transport) {
-        tripRouteOptionFindService.findTripRouteOptionCandidate(tripId, transport)
-                .ifPresent(routeOption -> {
-                    tripSegmentCommandService.deleteAllByRouteOption(routeOption);
-                    tripRouteOptionCommandService.delete(routeOption);
-                });
-    }
-
-    private List<TripSegmentEntity> createTripSegments(
-            TripRouteOptionEntity routeOption,
-            List<Section> sections,
-            List<TripPlaceResponse> tripPlaces
-    ) {
-        List<TripPlaceResponse> sortedTripPlaces = new ArrayList<>(tripPlaces);
-        sortedTripPlaces.sort(tripPlaceRouteOrder());
-
-        List<TripSegmentEntity> tripSegments = new ArrayList<>();
-        for (int i = 0; i < sortedTripPlaces.size() - 1 && i < sections.size(); i++) {
-            tripSegments.add(createTripSegment(routeOption, sections.get(i), sortedTripPlaces.get(i), sortedTripPlaces.get(i + 1)));
-        }
-
-        return tripSegments;
-    }
-
-    private TripSegmentEntity createTripSegment(
-            TripRouteOptionEntity routeOption,
-            Section section,
-            TripPlaceResponse startPlace,
-            TripPlaceResponse endPlace
-    ) {
-        TripPlaceEntity startTripPlace = tripPlaceService.getTripPlaceById(startPlace.getTripPlaceId());
-        TripPlaceEntity endTripPlace = tripPlaceService.getTripPlaceById(endPlace.getTripPlaceId());
-
-        return TripSegmentEntity.builder()
-                .routeOption(routeOption)
-                .startTripPlace(startTripPlace)
-                .endTripPlace(endTripPlace)
-                .distance(section.getDistance())
-                .duration(toDurationInMinutes(section))
-                .carbonEmission(section.getCarbonEmission())
-                .build();
-    }
-
-    @Transactional
     public DistanceResponse calculateAndSaveCarRoute(CarDistanceRequest request, Long tripId) {
         List<TripPlaceResponse> tripPlaces = tripPlaceService.getTripPlaces(tripId);
 
@@ -161,12 +52,11 @@ public class TripSegmentService {
         }
 
         List<Section> sections = carResponse.getSections();
-        saveRouteCandidate(tripId, sections, TripTransport.CAR.getCode(), tripPlaces);
+        tripRouteCandidateService.saveRouteCandidate(tripId, sections, TripTransport.CAR, tripPlaces);
 
         return carResponse.getDistanceResponse();
     }
 
-    @Transactional
     public List<RouteCalculationResult> calculateAndSavePublicRoute(TransDistanceBatchRequest batchRequest, List<TripPlaceResponse> tripPlaces) {
         List<RouteCalculationResult> results = transDistanceService.getBatchDistance(batchRequest);
 
@@ -180,7 +70,7 @@ public class TripSegmentService {
             sections.add(section);
         }
 
-        saveRouteCandidate(batchRequest.getTripId(), sections, TripTransport.PUBLIC.getCode(), tripPlaces);
+        tripRouteCandidateService.saveRouteCandidate(batchRequest.getTripId(), sections, TripTransport.PUBLIC, tripPlaces);
 
         return results;
     }
@@ -332,23 +222,4 @@ public class TripSegmentService {
                 .thenComparing(TripPlaceResponse::getVisitOrder);
     }
 
-    private TripRouteOptionEntity createRouteOptionFromSections(
-            TripEntity trip,
-            TripTransport transport,
-            List<Section> sections,
-            boolean confirmed
-    ) {
-        return TripRouteOptionEntity.builder()
-                .trip(trip)
-                .transport(transport)
-                .totalDistance(sections.stream().mapToDouble(Section::getDistance).sum())
-                .totalDuration(sections.stream().mapToInt(this::toDurationInMinutes).sum())
-                .totalCarbonEmission(sections.stream().mapToDouble(Section::getCarbonEmission).sum())
-                .confirmed(confirmed)
-                .build();
-    }
-
-    private int toDurationInMinutes(Section section) {
-        return Math.max(1, section.getDuration() / 60);
-    }
 }
